minor UI changes

+++ frontend/src/components/WorkoutDetails.js
@@ -7,13 +7,15 @@ const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutsContext();
 
   const handleClick = async () => {
-    const response = await fetch("/api/workouts/" + workout._id, {
-      method: "DELETE",
-    });
-    const json = await response.json();
+    if(window.confirm("Delete Workout?")){
+      const response = await fetch("/api/workouts/" + workout._id, {
+        method: "DELETE",
+      });
+      const json = await response.json();
 
-    if (response.ok) {
-      dispatch({ type: "DELETE_WORKOUT", payload: json });
+      if (response.ok) {
+        dispatch({ type: "DELETE_WORKOUT", payload: json });
+      }
     }
   };
   return (
